+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 CHANGELOG for Sulu Product Bundle
 =================================
 
+* dev-develop
+    * FEATURE     Added error handling when json-schema is not valid.
+
 * 0.1.0 (2016-11-02)
     * FEATURE     Added tests for get and post schema validation.
     * BUGFIX      Fixed validation issue, when no payload is provided.